Add a .ipynb file for Part 0
Preprocessing.ipynb file is for processing input feature vectors. I use numpy and pandas for dealing with matrix, and matplotlib for visualization. Please re-write a path to data when you use the local environment. Welcome to fix the code for improvement! If you have any questions, please let me know.