Implementation of linear regression
This is an implementation script of linear regression. Although the basic framework might be correct, there are some bugs so that the code does not work correctly. 
The problem is: the gradient does not converge but diverge so that I should fix the self.__gradient Method. Another concern is the training method. From the guideline, I had to set a convergence criterion to stop the training procedure. While this criterion is scaler value, nabla L is a vector. Should I calculate the scale of the vector then compare the scale and the criterion?